Now, we can open a demo window several times.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@12075 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/examples/vision2/test_all_widgets/main/main_window.e
@@ -26,8 +26,7 @@ creation
 
 feature --Access
 	
-	container: EV_VERTICAL_BOX
-			-- Push buttons
+	container: EV_DYNAMIC_TABLE
 	
 	current_demo_window: DEMO_WINDOW
 	
@@ -47,9 +46,14 @@ feature -- Initialization
 			c9: SCROLLABLE_AREA_DEMO_WINDOW
 			c10: BUTTONS_DEMO_WINDOW
 			c11: LIST_DEMO_WINDOW
+			c12: TABLE_DEMO_WINDOW
+			c13: DYNAMIC_TABLE_DEMO_WINDOW
+			c14: TREE_DEMO_WINDOW
 		do
 			Precursor
 			!!container.make (Current)
+			container.set_finite_dimension (2)
+			container.set_row_layout
 			!!c1.make (Current)
 			!!c2.make (Current)
 			!!c3.make (Current)
@@ -61,31 +65,37 @@ feature -- Initialization
 			!!c9.make (Current)
 			!!c10.make (Current)
 			!!c11.make (Current)
-			
+			!!c12.make (Current)		
+			!!c13.make (Current)
+			!!c14.make (Current)
 			!!b.make_button (Current, "Label", "", c1)
-			!!b.make_button (Current, "Buttons", "../pixmaps/buttons.xpm", c10)
-			!!b.make_button (Current, "Fixed", "../pixmaps/fixed.xpm", c2)
-			!!b.make_button (Current, "Box", "../pixmaps/box.xpm", c3)
-			!!b.make_button (Current, "Notebook", "../pixmaps/notebook.xpm", c4)
-			!!b.make_button (Current, "Text field", "../pixmaps/text_field.xpm", c5)
-			!!b.make_button (Current, "Text area", "../pixmaps/text_area.xpm", c6)
-			!!b.make_button (Current, "Menu", "../pixmaps/menu.xpm", c7)
-			!!b.make_button (Current, "Split area", "../pixmaps/split_area.xpm", c8)
-			!!b.make_button (Current, "Scrollable area", "../pixmaps/scrollable_area.xpm", c9)
+			!!b.make_button (Current, "Buttons", "pixmaps/buttons.xpm", c10)
+			!!b.make_button (Current, "Fixed", "pixmaps/fixed.xpm", c2)
+			!!b.make_button (Current, "Box", "pixmaps/box.xpm", c3)
+			!!b.make_button (Current, "Notebook", "pixmaps/notebook.xpm", c4)
+			!!b.make_button (Current, "Text field", "pixmaps/text_field.xpm", c5)
+			!!b.make_button (Current, "Text area", "pixmaps/text_area.xpm", c6)
+			!!b.make_button (Current, "Menu", "pixmaps/menu.xpm", c7)
+			!!b.make_button (Current, "Split area", "pixmaps/split_area.xpm", c8)
+			!!b.make_button (Current, "Scrollable area", "pixmaps/scrollable_area.xpm", c9)
 			!!b.make_button (Current, "List", "", c11)
+			!!b.make_button (Current, "Table", "pixmaps/table.xpm", c12)
+			!!b.make_button (Current, "Dynamic Table", "pixmaps/dtable.xpm", c13)
+			!!b.make_button (Current, "Tree", "pixmaps/tree.xpm", c14)
 
 			set_values
 		end
 	
 feature -- Status setting
 	
-
-	
 	execute (arg: EV_ARGUMENT1[DEMO_WINDOW]) is
 			-- called when actions window is deleted
 		do
- 			arg.first.effective_button.set_pressed (False)
-			arg.first.actions_window.destroy
+ 			arg.first.effective_button.set_state (False)
+			if arg.first.actions_window /= Void then
+				arg.first.actions_window.destroy
+			end
+			arg.first.hide
 			set_insensitive (False)
 		end
